+++ pytket/docs/changelog.md
@@ -5,6 +5,7 @@
 Fixes:
 
 - Update `Decompose_TK2` to not check decomposition with wire swaps when the `TK2` gate only has a single angle.
+- Make `optypeinfo()` thread-safe.
 
 
 ## 2.9.3 (August 2025)